# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample NODEJS server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
    "version": "1.0.0",
    "title": "Swagger NodeJS base",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:9000",
  "basePath": "/api",
...
  "schemes": ["http", "https"],
  "paths": {
...
    "/auth/refresh-token": {
      "post": {
        "tags": ["auth"],
        "summary": "Get new token",
        "description": "",
        "operationId": "refreshToken",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User information object that needs to be login",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "refresh_token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}